Changes to follow any re-directions and error checking
Added options to the needle.get to follow any re-directions because if the VIN is searched for the first time decodethis.com does some re-directions before it returns the response, so by setting "follow: true" fixes the issue. I have also enabled error checking by passing the "error" to the callback function or null if no error, so this will give more visibility to the user about the error if any.